backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
diffusers
transformers
accelerate
safetensors
pillow
opencv-python
